Fixed - get children call fails when one of the file name has special character (#5)
* Override method while returns FileObject for children.
It is overridden to prevent failing call for returning children when one of the file has invalid character in name (i.e. test%test.mp4)

* Added GCSFileName and GCSFileNameParser to handle special characters in file name.
Earlier it was using URLFileNameParser and UrlFileName however it fails to decode the character like % when they are in file name (i.e. test%test.mp4_

* Renamed new classes as per java convention

* Renamed new classes as per java convention

* Renamed variable name to pathStringBuilder

Co-authored-by: Ashok Mor <amor@dalet.com>